Create SingleSelectQuestionTest
